1. 增加计算器，会弹出一个新的窗口； 2.打印页面客户规格前会自动带上厚度。3. 入库出库去掉吨的小数位限制，4. 订单和销售页面，如果选张会吧数量的精度改为1，选吨会把精度改为3

+++ src/main/dbscripts/steel.sql
@@ -10,8 +10,8 @@ CREATE TABLE
     (
         inventory_id INT NOT NULL AUTO_INCREMENT,
         inventory_date DATE,
-        store_in DOUBLE(10,4) DEFAULT 0.0000,
-        store_out DOUBLE(10,4) DEFAULT 0.0000,
+        store_in double(20,10) DEFAULT 0.0000,
+        store_out double(20,10) DEFAULT 0.0000,
         YEAR INT,
         MONTH VARCHAR(200),
         steel_name VARCHAR(200),
@@ -24,7 +24,7 @@ DROP TABLE steel_order;
 CREATE TABLE steel_order (order_id int NOT NULL AUTO_INCREMENT, order_date date not null, order_no varchar(45) not null, client_id int not null, account_no varchar(45) not null, spec_id int  not null,client_spec varchar(100) not null, client_amount double(10,4)  not null, price double(10,4)  not null, cash_amount double(16,4)  not null, steel_calc_amount varchar(100)  not null, comment varchar(400), is_out int(2)  not null COMMENT '0-未出库  1-已出库', is_sale int(2)   not null COMMENT '0-未销售  1-已销售', is_delete int(2) not null, unit varchar(45)  not null, year int(4)  not null, month int(2)  not null, PRIMARY KEY (order_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci COMMENT='钢板订单表';
 
 DROP TABLE steel_outbound;
-CREATE TABLE steel_outbound (out_id int NOT NULL AUTO_INCREMENT, order_no varchar(45), spec_id int, actual_amount double(10,4), out_date date, year int, month int, PRIMARY KEY (out_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci;
+CREATE TABLE steel_outbound (out_id int NOT NULL AUTO_INCREMENT, order_no varchar(45), spec_id int, actual_amount double(20,10), out_date date, year int, month int, PRIMARY KEY (out_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci;
 ALTER  TABLE  `steel_outbound`  ADD  UNIQUE (`order_no` );
 
 DROP TABLE steel_parameter;
@@ -41,7 +41,7 @@ DROP TABLE steel_specs;
 CREATE TABLE steel_specs (spec_id int NOT NULL, thickness double(10,3), steel_code varchar(45), spec_desc varchar(300), category_id int, price_code varchar(45), PRIMARY KEY (spec_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci COMMENT='钢板规格信息，主要维护钢板厚度';
 
 DROP TABLE steel_storage;
-CREATE TABLE steel_storage (storage_id int NOT NULL AUTO_INCREMENT, store_date date, client_no varchar(100), store_no varchar(100), steel_amount double(10,4), steel_factory varchar(200), year int, month int, spec_id int, client_id int, cash_amount double(10,2), price double(10,4), PRIMARY KEY (storage_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci COMMENT='工厂钢板入库信息表';
+CREATE TABLE steel_storage (storage_id int NOT NULL AUTO_INCREMENT, store_date date, client_no varchar(100), store_no varchar(100), steel_amount double(20,10), steel_factory varchar(200), year int, month int, spec_id int, client_id int, cash_amount double(10,2), price double(10,4), PRIMARY KEY (storage_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci COMMENT='工厂钢板入库信息表';
 
 DROP TABLE user;
 CREATE TABLE user (user_id int NOT NULL, user_name varchar(45), password varchar(45), role varchar(45), PRIMARY KEY (user_id)) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE=utf8_general_ci;
